Weather_App_Project V1.1 After Solving Errors
first problem 
Post route
Implement async calling by the use of promise chaining where you will pass the the mix of API and user responses, to POST endpoint on server side.
The POST route in the server side should setup which will map the data sent via the API from the client side and save it in the projectData variable.

Second problem 
return endpoint data 
get route 1:server side 
There should be a GET route setup on the server side with the first argument as a string naming the route, and the second argument a callback function to return the JS object created at the top of server code.

third problem 
Integrating openweathermap API 
The personal API Key for OpenWeatherMap API is saved in a named const variable. The API Key variable is passed as a parameter to fetch(). Data is successfully returned from the external API.

In the file app.js, there should be a line of code near the top:

forth problem 
return endpoint data 
get route 2:client side
There should be an asynchronous function to fetch the data from the app endpoint


fifth problem 
Dynamicaly update UI 
Sets the properties of existing HTML elements from the DOM using Vanilla JavaScript.

Included in the async function to retrieve that app’s data on the client side, existing DOM elements should have their innerHTML properties dynamically set according to data returned by the app route.

sixth problem 
event listeners 
Adds an event listener to an existing HTML button from DOM using Vanilla JS.

In the file app.js, the element with the id of generate should have an addEventListener() method called on it, with click as the first parameter, and a named callback function as the second parameter.